separate file QuickInputFab, proper desc tooltip
